requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
apturl==0.5.2
art==6.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.4
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
einops==0.7.0
fasteners==0.14.1
filelock==3.9.0
flash-attn==2.5.5
flatbuffers==24.3.25
frozendict==2.4.1
frozenlist==1.4.1
fsspec==2024.3.1
future==0.18.2
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
langchain==0.1.9
langchain-community==0.0.32
langchain-core==0.1.41
langsmith==0.1.45
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==5.2.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.21.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
netifaces==0.11.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.17.3
orjson==3.10.0
packaging==23.2
pandas==2.2.0
paramiko==2.9.3
peewee==3.17.1
pexpect==4.8.0
Pillow==9.0.1
piper-phonemize==1.1.0
piper-tts==1.2.0
pocketsphinx==5.0.3
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
PyAudio==0.2.14
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.6.2
pydantic_core==2.16.3
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.34.0
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
sympy==1.12
systemd-python==234
tenacity==8.2.3
tokenizers==0.15.2
torch==2.1.2
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
transformers==4.39.3
triton==2.1.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
yarl==1.9.4
yfinance==0.2.36
zipp==1.0.0
